Add alternative bypass_cache
Instead of having bypass_cache as a Query object attribute, see if it is more palatable as query parameter.
Is it possible to have query parameter alongside a requestBody? I assume so, but I don't know this to work
For the `extended` branch
The assumption is that if a server does not implement this, then there is no caching. If you implement caching, you should do it this way.