Update resource to resource_id
This is a proposed CHANGE to TRAPI 1.4-beta, to be included in ongoing implementations. As discussed in the TRAPI call today, this `resource` and `upstream_resources` go against our convention that such properties that are a CURIE should end in _id. Rename these to `resource_id` and `upstream_resource_ids`